CE: moved getSocialLoginButtonProps func to EE file as it's dependencies will be updated in EE
