+++ app/client/src/ce/constants/SocialLogin.tsx
@@ -29,15 +29,3 @@ export const SocialLoginButtonPropsList: Record<
 };
 
 export type SocialLoginType = keyof typeof SocialLoginButtonPropsList;
-
-export const getSocialLoginButtonProps = (
-  logins: SocialLoginType[],
-): SocialLoginButtonProps[] => {
-  return logins.map((login) => {
-    const socialLoginButtonProps = SocialLoginButtonPropsList[login];
-    if (!socialLoginButtonProps) {
-      throw Error("Social login not registered: " + login);
-    }
-    return socialLoginButtonProps;
-  });
-};
